split code into several classes
