+++ src/main/java/ru/mercuriev/game/of/life/no/condition/NeighboursCollector.java
@@ -1,6 +1,8 @@
 package ru.mercuriev.game.of.life.no.condition;
 
 import ru.mercuriev.game.of.life.no.condition.cells.Cell;
+import ru.mercuriev.game.of.life.no.condition.worlds.World;
+import ru.mercuriev.game.of.life.no.condition.worlds.WrappedWorld;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,11 +11,18 @@ public class NeighboursCollector {
 
     WrappedWorld wrappedWorld;
 
-    public NeighboursCollector(WrappedWorld wrappedWorld) {
-        this.wrappedWorld = wrappedWorld;
+    public NeighboursCollector(World world) {
+        this.wrappedWorld = WrappedWorld.newInstance(world);
     }
 
-    public List<Cell> collect(Cell cell) {
+    public int sum(Cell c) {
+        return this.collect(c)
+                .stream()
+                .mapToInt(Cell::getState)
+                .sum();
+    }
+
+    List<Cell> collect(Cell cell) {
         List<Cell> n = new ArrayList<>();
 
         int i = cell.getX();
@@ -31,12 +40,5 @@ public class NeighboursCollector {
         n.add(wrappedWorld.cellAt(i + 1, j + 1));
 
         return n;
-    }
-
-    public int sum(Cell c) {
-        return this.collect(c)
-                .stream()
-                .mapToInt(Cell::getState)
-                .sum();
     }
 }
