+++ src/main/java/ru/mercuriev/game/of/life/no/condition/NoConditionEngine.java
@@ -1,9 +1,8 @@
 package ru.mercuriev.game.of.life.no.condition;
 
 import org.springframework.stereotype.Service;
-import ru.mercuriev.game.of.life.no.condition.cells.Cell;
+import ru.mercuriev.game.of.life.no.condition.worlds.World;
 
-import java.util.List;
 import java.util.stream.Collectors;
 
 import static ru.mercuriev.game.of.life.no.condition.DisionMaker.nextGenCell;
@@ -11,19 +10,21 @@ import static ru.mercuriev.game.of.life.no.condition.DisionMaker.nextGenCell;
 @Service
 public class NoConditionEngine {
 
-    public int[][] next(int[][] ints) {
+    public int[][] next(int[][] world) {
+        World world1 = World.newInstance(world);
 
-        World world = World.newInstance(ints);
-        WrappedWorld w = WrappedWorld.newInstance(world);
-        NeighboursCollector neighboursCollector = new NeighboursCollector(w);
+        NeighboursCollector neighboursCollector = new NeighboursCollector(world1);
 
-        List<List<Cell>> nextGeneration = world.getRows().map(row ->
-                row.stream()
-                        .map(cell -> nextGenCell(cell, neighboursCollector.sum(cell)))
-                        .collect(Collectors.toList())
-        ).collect(Collectors.toList());
+        World nextGenWorld = World.fromList(
+                world1.getRows().map(row ->
+                        row.stream()
+                                .map(cell -> nextGenCell(cell, neighboursCollector.sum(cell)))
+                                .collect(Collectors.toList())
+                ).collect(Collectors.toList())
+        );
 
-        return World.newInstance(nextGeneration).toArray();
+        return nextGenWorld.toArray();
     }
+
 
 }
