+++ src/test/java/ru/mercuriev/game/of/life/no/condition/DisionMakerTest.java
@@ -3,6 +3,8 @@ package ru.mercuriev.game.of.life.no.condition;
 import org.junit.Test;
 import ru.mercuriev.Fixture;
 import ru.mercuriev.game.of.life.no.condition.cells.Cell;
+import ru.mercuriev.game.of.life.no.condition.worlds.World;
+import ru.mercuriev.game.of.life.no.condition.worlds.WrappedWorld;
 
 import java.util.stream.IntStream;
