+++ src/test/java/ru/mercuriev/game/of/life/no/condition/WrappedWorldTest.java
@@ -2,6 +2,8 @@ package ru.mercuriev.game.of.life.no.condition;
 
 import org.junit.Test;
 import ru.mercuriev.Fixture;
+import ru.mercuriev.game.of.life.no.condition.worlds.World;
+import ru.mercuriev.game.of.life.no.condition.worlds.WrappedWorld;
 
 import java.util.stream.IntStream;
